# Upcasting, scaling, masking and fused kernels to match Megatron-LM