--- a/src/transformers/models/gpt_bigcode/configuration_gpt_bigcode.py
+++ b/src/transformers/models/gpt_bigcode/configuration_gpt_bigcode.py
@@ -166,7 +166,7 @@ class GPTBigCodeConfig(PretrainedConfig):
         bos_token_id=50256,
         eos_token_id=50256,
         attention_softmax_in_fp32=False,
-        scale_attention_softmax_in_fp32=False,
+        scale_attention_softmax_in_fp32=True,
         attention_type=AttentionType.MULTI_HEAD,
         **kwargs,
     ):