Run the following sequence (when enabled):
* attn_weights = (query x key) / sqrt(head_dim) / (layer_idx+1) [Single kernel, I think that's needed to stabilize fp16]
* Convert to fp32
* multiply by (layer_idx+1)
* Apply mask
* Softmax
* Convert to fp16

The whole thing runs in 2 kernels with jit fusion. Cuda time inference benchmark:
* Without jit: 793.142ms
* With jit: 751.869ms

Without upcast, the sequence is:
Run the following sequence (when enabled):
* attn_weights = (query x key) / sqrt(head_dim)
* Apply mask
* Softmax

Jit can reduce it to 2 kernels:
* Without jit: 745.253ms
* With jit: 751.365ms

Jit looks slower so I disabled it.

Overall, jit fusion allows fp32 softmax with minimal overhead (0.9% instead of 6.4%) and without the complicated custom kernels from Megatron.